# Improve performance of EmsFolders and EmsClusters child save blocks