@agrare 2018-05-15
I suspect the extra db lookups in the graph refresh custom save blocks are the culprit for the slower performance on graph refresh.  I am going to try replacing those with lazy_finds to hopefully save a db lookup if we have the id cached already.